+++ sarc/users/__init__.py
@@ -1,0 +1,1 @@
+from . import mila_ldap  # noqa: F401 # This is to register the builtin plugins
